login handler
needs to be linked to the app page but both post/get should be working.
currently does not return anything and just redirects on itself for
errors

may be minor index errors from python complier but their shouldn't be
any tabs

side comment: because we are choosing to index by student pin number
this causes the weird gql interface needed.