Fixed merge issues.
I spent over 3 hours trying to fix all of the mistakes. Some were small, some were large. In addition to taking over the responsiblities of the local portion since one of our members is MIA, I also had to correct all of the issues with the Interface and Server codes. To be honest, I'm not sure if either were tested in-browser. I ran into several errors from code that would just not compile.