Additional interface tweaks.
Please note that these are not actually my responsibility. The tasks as assigned were not completed on time therefore, in order to meet our meeting deadline, I had to do an extensive amount of work to make the interface work.

In addition, it appears that several contributors are not utilizing the git PULL command before commencing work. When I tried to merge changes from the interface branch, the branch was 12 commits behinds master. I had to put a substantial amount of effort into fixing the merge conflicts because it was not updated. I think many of the contributors are confused as to how Git works.